refactor: data 패키지 구조 변경
- data 패키지 내부에 local, repository 패키지를 추가하여 계층을 분리했습니다.
  - local: Room DB 관련 클래스 (entity, dao, database)
  - repository: 데이터 접근 로직을 담당하는 Repository 인터페이스 및 구현체
