+++ app/src/main/java/com/example/myroutine/data/local/dao/RoutineDao.kt
@@ -1,9 +1,10 @@
-package com.example.myroutine.data
+package com.example.myroutine.data.local.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import com.example.myroutine.data.local.entity.RoutineItem
 
 @Dao
 interface RoutineDao {
