+++ app/src/main/java/com/example/myroutine/data/local/database/AppDatabase.kt
@@ -1,8 +1,12 @@
-package com.example.myroutine.data
+package com.example.myroutine.data.local.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import com.example.myroutine.data.local.dao.RoutineCheckDao
+import com.example.myroutine.data.local.dao.RoutineDao
+import com.example.myroutine.data.local.entity.RoutineCheck
+import com.example.myroutine.data.local.entity.RoutineItem
 
 @Database(entities = [RoutineItem::class, RoutineCheck::class], version = 1)
 @TypeConverters(Converters::class)
