+++ app/src/main/java/com/example/myroutine/data/local/entity/RoutineCheck.kt
@@ -1,4 +1,4 @@
-package com.example.myroutine.data
+package com.example.myroutine.data.local.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
