+++ app/src/main/java/com/example/myroutine/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.example.myroutine.di
 
-import com.example.myroutine.data.RoutineRepository
-import com.example.myroutine.data.RoutineRepositoryImpl
+import com.example.myroutine.data.repository.RoutineRepository
+import com.example.myroutine.data.repository.RoutineRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
